% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datainput-module.R
\name{dataInputModule}
\alias{dataInputModule}
\title{Server part of pavian data input module.}
\usage{
dataInputModule(
  input,
  output,
  session,
  config_dir = NULL,
  server_access = getOption("pavian.server_access", default = FALSE),
  load_server_directory = getOption("pavian.load_server_directory", default = FALSE),
  load_example_data = getOption("pavian.load_example_data", default = FALSE),
  pavian_options = NULL,
  db_type = getOption("pavian.db_type"),
  db_name = getOption("pavian.db_name"),
  db_host = getOption("pavian.db_host"),
  db_port = getOption("pavian.db_port"),
  db_user = getOption("pavian.db_user"),
  db_passwd = getOption("pavian.db_passwd")
)
}
\arguments{
\item{input}{Scoped input.}

\item{output}{Module output.}

\item{session}{Shiny session.}

\item{config_dir}{Directory for configuration files.}

\item{server_access}{Whether a directory on the server can be loaded.}

\item{load_server_directory}{Load server directory.}

\item{load_example_data}{Load example data.}

\item{pavian_options}{General options for pavian.}
}
\value{
Shiny module server function, to be called by callModule.
}
\description{
Server part of pavian data input module.
}
